+++ src/main/java/com/travelservice/domain/product/repository/ImageRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.travelservice.domain.product.entity.ProductImage;
 
 @Repository
-public interface ImageRepository extends JpaRepository<ProductImage, Integer> {
+public interface ImageRepository extends JpaRepository<ProductImage, Long> {
 }
